requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.5
sqlalchemy==1.4.41
boto3==1.26.47
pytesseract==0.3.10
PyPDF2==3.0.0
gradio
pytest==7.2.0
StrEnum==0.4.15
six==1.16.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
huggingface-hub==0.20.3
onnxruntime-gpu==1.17.1
StrEnum
pdfplumber
fasttext-wheel
datrie
nltk
torch
xgboost
shapely
pyclipper
hanziconv